Moved ApiForm and StatusIcon.vue into `api`
