+++ frontend/src/components/form/api/ApiSelect.vue
@@ -25,7 +25,7 @@ Displays a field as a v-select + write access via API wrapper
 <script>
 import { apiPropsMixin } from '@/mixins/apiPropsMixin'
 import ApiWrapper from './ApiWrapper'
-import StatusIcon from '../base/StatusIcon'
+import StatusIcon from './StatusIcon'
 
 export default {
   name: 'ApiSelect',
